UI + Settings cleanup. Added some additional mixins.

+++ src/main/java/net/aoba/module/modules/combat/Reach.java
@@ -62,6 +62,6 @@ public class Reach extends Module {
 	}
 	
 	public void setReachLength(float reach) {
-		this.distance.setValue((double)reach);
+		this.distance.setValue(reach);
 	}
 }
