+++ src/main/java/net/aoba/module/modules/movement/Freecam.java
@@ -50,7 +50,7 @@ public class Freecam extends Module implements TickListener {
 	}
 
 	public void setSpeed(float speed) {
-		this.flySpeed.setValue((double)speed);
+		this.flySpeed.setValue(speed);
 	}
 	
 	public double getSpeed() {
