+++ src/main/java/net/aoba/module/modules/world/AutoFarm.java
@@ -60,7 +60,7 @@ public class AutoFarm extends Module implements TickListener {
 	}
 
 	public void setRadius(int radius) {
-		this.radius.setValue((double)radius);
+		this.radius.setValue((float)radius);
 	}
 
 	@Override
